made new migrations, and add UserProfile data returned with User data
