+++ backend/account/urls.py
@@ -2,5 +2,6 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('register/', views.register, name='register')
+    path('register/', views.register, name='register'),
+    path('me/', views.current_user, name='current_user'),
 ]
