+++ backend/account/views.py
@@ -1,11 +1,13 @@
 from django.shortcuts import render
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
 
 from django.contrib.auth.hashers import make_password
 from .serializers import SignUpSerializer, UserSerializer
 from django.contrib.auth.models import User
+
+from rest_framework.permissions import IsAuthenticated
 
 # Create your views here.
 
@@ -38,3 +40,10 @@ def register(request):
         
     else:
         return Response(user.errors)
+    
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def current_user(request):
+    user = UserSerializer(request.user)
+    
+    return Response(user.data)
