+++ backend/music/serializers.py
@@ -1,14 +1,7 @@
 from rest_framework import serializers
-from .models import UserProfile, Message
+from .models import Message
 from django.contrib.auth.models import User
 
-
-
-class UserProfileSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = UserProfile
-        # fields to return?
-        fields = "__all__"
 
 class MusicMessageSerializer(serializers.ModelSerializer):
     class Meta:
